[maui] Moved around authroization logic, new static helper class
